+++ src/targetWidths.js
@@ -6,12 +6,12 @@ function targetWidths() {
 
   const ensureEven = n => 2 * Math.round(n / 2);
 
-  resolutions.push(prev);
   while (prev <= MAX_SIZE) {
-    prev *= 1 + (INCREMENT_PERCENTAGE / 100) * 2;
     resolutions.push(ensureEven(prev));
+    prev *= 1 + (INCREMENT_PERCENTAGE / 100) * 2;
   }
 
+  resolutions.push(MAX_SIZE);
   return resolutions;
 }